Initial port to the CMake Build System
This is a initial port, it is by no means finished. currently
it supports building the application using dynamic libraries.

I had to disable translations for the time being, which made me
realize that all strings in the application are in portuguese -
this should be translated to english.

to compile the software:
cd br-print
mkdir build
cmake ..
make

If you wanna create a debug build: cmake .. -DCMAKE_BUILD_TYPE=Debug

Signed-off-by: Tomaz Canabrava <tomaz.canabrava@intel.com>