Arduino Thread
This files start a Thread listener to find the port at the
Arduino is connect.
On previous versions submit to git, this thread is disabled,
because is not work with others Arduinos that no made by the
Italy Company, because on file that i search those hardwares
dont have the 'Arduino' word in the name. But after a test,
was realize that the thread is really needed. Because the ports
that the app show is not the only that the Arduino could connect.
So show up the default ports and new ports if the Arduino was
connect on then. Just works if the Arduino is original.

Signed-off-by: Lays Rodrigues <laysrodriguessilva@gmail.com>