Fixed the cmake to be used with older versions
The old cmake did by tomaz was using his own cmake version, 3.2
and debian only has 3.0, since the cmake changes where needed
for me to compile it, and since I didn't had to change anything
on the cmake code for it, I'm downgrading the version.

Also, add a bit of VTK library to test later. probably is a wrong
commit but we can always ammend that.

Signed-off-by: Lays Rodrigues <laysrodriguessilva@gmail.com>