Add new Methods
-StartPrintJob -> to start the print job.
-SetPrintLog -> this function is needed because when the
print job start, the user can set if he wants a log or not.
And i need catch the status of printLog on PrinterSettings
Widget. So when the user click on the checkBox emit a signal
to this class and update a boolean variable to the current
status of the checkBox PrintLog.

Signed-off-by: Lays Rodrigues <laysrodriguessilva@gmail.com>