Fix Slot
In the slot isPrintJob running, i need to know if the
pause button is checked or not. So before I would need
a get function to know what the status of the button.
However, i have the slot to stop the print job that receive
the boolean that i need to know. So i just made a local
variable receive the status of the button.

Signed-off-by: Lays Rodrigues <laysrodriguessilva@gmail.com>