Remove slots
The start/stop ThreadRoutine was remove, because I realize
that would be better use a QTimer instead of QThread.
The change was made to reduce the amount of resources used,
since a thread has considerable use of memory and cpu.
Also the change, is a attempt to reduce the quantity
of the DeadLocks that is occurring.

Signed-off-by: Lays Rodrigues <laysrodriguessilva@gmail.com>