Improve Method
Add to the method renderGcode the separation of
car moves and print moves, and render those separate.
Only allows the strings of gcode of size that i need to know
about this informations.
I need to analyze how to fix the read of point Z when the
string analyze have size of 5, because the first if to read
X and Y points only work on the second iteration of the for,
since the test of Z work on the two iterations, the algorithm
read the Z point two times and add more layers that the gcode
have.

Signed-off-by: Lays Rodrigues <laysrodriguessilva@gmail.com>