Updates
Update the methods that was made to work with QLineEdit
to work with the QSpinBox.
Also update de connections used by the QSpinBox
Update the method sendValue to dont receive any parameters
and consult the data need on the slot. Because the use of
signal QSpinBox::valueChange(int) use some things about
C++ and static cash that i dont know.

Signed-off-by: Lays Rodrigues <laysrodriguessilva@gmail.com>