Start use RegularExpressions
On the method readGcode, we need support two kinds
of gcodes, to cartesians and delta printers. The
previous method only read gcode of carteasian printers.
So the method now is use regular expressions to read the
values from gcode.
This method isnt finish yet. Because isnt loading
the gcode of delta printers, i think that the cause
is the regular expression. But i will figure out soon.
-Add the regular expressions and made the ajustments to
use it.

Signed-off-by: Lays Rodrigues <laysrodriguessilva@gmail.com>