Ajustments
Well, I will need use 5 sliders on the ManualControlWidget, to
manage settings of the 3dprinter. 2 sliders, will need have
some actions from a button, so made the ajustments using a boolean
indicating if I will need or not use a QPushButton.
Made the includes and ajusts necessary to work with
QGraphicsPixmapItem or QPushButton.

Signed-off-by: Lays Rodrigues <laysrodriguessilva@gmail.com>