New Class
This class is a widget to manage extruders
that the 3dprinter could have, if the extruderInUse > 2.
I though that is more easy to the user have one place
to managem them all, with a few tools that he needs
to do that.
Create some methods and made the necessary adjusts for a
simple use.

Signed-off-by: Lays Rodrigues <laysrodriguessilva@gmail.com>