Change app to support only ONE extruder
Several changes to change the support of the qnt of extruders
in this app.
Since we will made a new version of this app, this version will
not receive new adjustments, like controllers to manage more
the one extruder.
To simplify this version, I remove all support for 2
extruders or more. Because the previous logic dont have
full support to more then one extruder.

Signed-off-by: Lays Rodrigues <laysrodriguessilva@gmail.com>